+++ osdcloud-assets/OSDCloud/WinPE/OSDCloud/Config/Scripts/Shutdown/Invoke-OobeCustomization.ps1
@@ -47,6 +47,7 @@ function Get-SelectedOsMetadata {
     }
 
     [pscustomobject]@{
+        uiLanguage = 'zh-TW'
         locale = 'zh-TW'
         language = 'zh-tw'
         timeZone = 'Taipei Standard Time'
@@ -117,15 +118,21 @@ try {
     Write-Host "Target Windows root: $windowsRoot"
     $deploymentMetadata = Get-DeploymentMetadata -WindowsRoot $windowsRoot
     $selectedOs = Get-SelectedOsMetadata -DeploymentMetadata $deploymentMetadata
-    $locale = if ($selectedOs.locale) { [string] $selectedOs.locale } elseif ($selectedOs.language) { [string] $selectedOs.language } else { 'zh-TW' }
-    $timeZone = if ($selectedOs.timeZone) { [string] $selectedOs.timeZone } else { 'Taipei Standard Time' }
+    $uiLanguage = if ($selectedOs.uiLanguage) { [string] $selectedOs.uiLanguage } elseif ($selectedOs.language) { [string] $selectedOs.language } else { 'zh-TW' }
+    $locale = if ($selectedOs.locale) { [string] $selectedOs.locale } else { $uiLanguage }
+    $timeZone = if ($selectedOs.timeZone) { [string] $selectedOs.timeZone } else { '' }
+    if ([string]::IsNullOrWhiteSpace($timeZone)) {
+        throw 'Deployment metadata is missing an explicit Windows time zone.'
+    }
     $localeXml = ConvertTo-XmlText -Value $locale
+    $uiLanguageXml = ConvertTo-XmlText -Value $uiLanguage
     $timeZoneXml = ConvertTo-XmlText -Value $timeZone
     $windowsUsername = Get-DeploymentSecret -JsonName 'windowsUsername' -EnvironmentName 'OSDCLOUD_WINDOWS_USERNAME'
     $windowsPassword = Get-DeploymentSecret -JsonName 'windowsPassword' -EnvironmentName 'OSDCLOUD_WINDOWS_PASSWORD'
     $windowsUsernameXml = ConvertTo-XmlText -Value $windowsUsername
     $windowsPasswordXml = ConvertTo-XmlText -Value $windowsPassword
-    Write-Host "OOBE locale: $locale"
+    Write-Host "OOBE UILanguage: $uiLanguage"
+    Write-Host "OOBE regional format (UserLocale): $locale"
     Write-Host "OOBE time zone: $timeZone"
 
     $panther = Join-Path $windowsRoot 'Windows\Panther'
@@ -138,9 +145,9 @@ try {
 <unattend xmlns="urn:schemas-microsoft-com:unattend">
   <settings pass="oobeSystem">
     <component name="Microsoft-Windows-International-Core" processorArchitecture="amd64" publicKeyToken="31bf3856ad364e35" language="neutral" versionScope="nonSxS" xmlns:wcm="http://schemas.microsoft.com/WMIConfig/2002/State" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-      <InputLocale>$localeXml</InputLocale>
-      <SystemLocale>$localeXml</SystemLocale>
-      <UILanguage>$localeXml</UILanguage>
+      <SystemLocale>$uiLanguageXml</SystemLocale>
+      <UILanguage>$uiLanguageXml</UILanguage>
+      <UILanguageFallback>$uiLanguageXml</UILanguageFallback>
       <UserLocale>$localeXml</UserLocale>
     </component>
     <component name="Microsoft-Windows-Shell-Setup" processorArchitecture="amd64" publicKeyToken="31bf3856ad364e35" language="neutral" versionScope="nonSxS" xmlns:wcm="http://schemas.microsoft.com/WMIConfig/2002/State" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">